#1339 - Move out migrations component into separate composer package
